api/requirements.txt
flatlib
pyswisseph
